Crush core/overdrive.py and api/overdrive.py together.
