+++ bin/informational/overdrive-advantage-accounts
@@ -6,6 +6,6 @@ bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..", "..")
 sys.path.append(os.path.abspath(package_dir))
 
-from core.scripts import GenerateOverdriveAdvantageAccountList
+from api.overdrive import GenerateOverdriveAdvantageAccountList
 
 GenerateOverdriveAdvantageAccountList().run()
