+++ bin/informational/overdrive-advantage-list
@@ -6,6 +6,6 @@ bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..", "..")
 sys.path.append(os.path.abspath(package_dir))
 
-from api.overdrive import OverdriveAdvantageAccountListScript  # noqa: E402
+from api.overdrive import OverdriveAdvantageAccountListScript
 
 OverdriveAdvantageAccountListScript().run()
