+++ bin/repair/overdrive_bibliographic_coverage
@@ -6,7 +6,7 @@ import sys
 bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..", "..")
 sys.path.append(os.path.abspath(package_dir))
-from core.overdrive import OverdriveBibliographicCoverageProvider
+from api.overdrive import OverdriveBibliographicCoverageProvider
 from core.scripts import RunCoverageProviderScript
 
 RunCoverageProviderScript(OverdriveBibliographicCoverageProvider).run()
