--- tests/core/files/overdrive/overdrive_availability_not_found.json
@@ -1,5 +1,0 @@
-{
-    "errorCode": "NotFound", 
-    "message": "The requested resource could not be found.", 
-    "token": "60a18218-0d25-42b8-80c3-0bf9df782f1b"
-}
